requirements.txt
python-telegram-bot==20.8
aiohttp>=3.8.1
requests>=2.25.1
